+++ src/main/kotlin/com/knu/mockin/exception/ExceptionDto.kt
@@ -1,4 +1,4 @@
-package com.knu.mockin.exeption
+package com.knu.mockin.exception
 
 data class ExceptionDto(
     val status: Int,
