+++ src/main/kotlin/com/knu/mockin/security/handler/JwtAccessDeniedHandler.kt
@@ -1,8 +1,7 @@
 package com.knu.mockin.security.handler
 
-import com.knu.mockin.exeption.CustomException
-import com.knu.mockin.exeption.ErrorCode
-import com.knu.mockin.exeption.ExceptionDto
+import com.knu.mockin.exception.ErrorCode
+import com.knu.mockin.exception.ExceptionDto
 import com.knu.mockin.logging.utils.LogUtil.toJson
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType.APPLICATION_JSON
