+++ src/main/kotlin/com/knu/mockin/service/AccountService.kt
@@ -1,6 +1,6 @@
 package com.knu.mockin.service
 
-import com.knu.mockin.exeption.ErrorCode
+import com.knu.mockin.exception.ErrorCode
 import com.knu.mockin.model.dto.request.account.KeyPairRequestDto
 import com.knu.mockin.model.dto.request.account.UserAccountNumberRequestDto
 import com.knu.mockin.model.dto.response.SimpleMessageResponseDto
