+++ src/main/kotlin/com/knu/mockin/util/ExtensionUtil.kt
@@ -1,7 +1,7 @@
 package com.knu.mockin.util
 
-import com.knu.mockin.exeption.CustomException
-import com.knu.mockin.exeption.ErrorCode
+import com.knu.mockin.exception.CustomException
+import com.knu.mockin.exception.ErrorCode
 import reactor.core.publisher.Mono
 
 object ExtensionUtil {
