+++ src/test/kotlin/com/knu/mockin/kisclient/KISTradingClientTest.kt
@@ -3,8 +3,8 @@ package com.knu.mockin.kisclient
 import com.knu.mockin.config.ConstantConfig
 import com.knu.mockin.dsl.RestDocsUtils
 import com.knu.mockin.dsl.toDto
-import com.knu.mockin.exeption.CustomException
-import com.knu.mockin.exeption.ErrorCode
+import com.knu.mockin.exception.CustomException
+import com.knu.mockin.exception.ErrorCode
 import com.knu.mockin.model.dto.request.trading.*
 import com.knu.mockin.model.enum.TradeId
 import com.knu.mockin.repository.UserRepository
